requirements.txt
pymongo
PyQt5
python-dotenv